+++ EssentialFeediOS/Feed UI/Composers/FeedUIComposer.swift
@@ -61,29 +61,3 @@ private final class FeedViewAdapter: FeedView {
         }
     }
 }
-
-private final class FeedLoaderPresentationAdaapter: FeedViewControllerDelegate {
-    private let feedLoader: FeedLoader
-    var presenter: FeedPresenter?
-
-    init(feedLoader: FeedLoader) {
-        self.feedLoader = feedLoader
-    }
-
-    func didRequestFeedRefresh() {
-        loadFeed()
-    }
-
-    private func loadFeed() {
-        presenter?.didStartLoadingFeed()
-        feedLoader.load { [weak self] result in
-            switch result {
-                case let .success(feed):
-                    self?.presenter?.didFinishLoadingFeed(with: feed)
-
-                case let .failure(error):
-                    self?.presenter?.didFinishLoadingFeed(with: error)
-            }
-        }
-    }
-}
